AD-34. UI. Training mode grid - Fixed FloorGrid - implemented X/Y cells count instead of X/Y size; - implemented RallyPoints; - fixed unit orientation when spawned; - added <Training> menu option.

+++ Source/AtomDestiny/Parameters/ParametersBase.cpp
@@ -69,7 +69,7 @@ void UParametersBase::InitializeComponent()
     m_currentHealth = m_maxHealth;
     m_currentMaxHealth = m_currentHealth;
     m_currentDefence = m_defence;
-    
+
     AddNewParameter(EObjectParameters::Health);
     AddNewParameter(EObjectParameters::MaxHealth);
     AddNewParameter(EObjectParameters::Defence);
@@ -78,7 +78,7 @@ void UParametersBase::InitializeComponent()
 void UParametersBase::BeginPlay()
 {
     Super::BeginPlay();
-    
+
     m_balanceParameters = AtomDestiny::Balance::CorrectBalanceParameters(m_balanceParameters);
 
     LOG_CHECK_WARNING(m_healthBarWidget != nullptr, TEXT("[ObjectStateBase::BeginPlay] HealthBar is NULL"));
@@ -109,7 +109,7 @@ double UParametersBase::GetDamageAfterDefenceParameters(EWeaponType type, double
     default:
         break;
     }
-    
+
     return damage - calculateResistanceValue;
 }
 
@@ -118,15 +118,15 @@ double UParametersBase::GetDamageAfterDefenceType(float damage) const
     switch (m_defenceType)
     {
     case EDefenceType::Light:
-        damage *= AtomDestiny::Balance::VsLightDefenceCoefficient;
+        damage *= static_cast<float>(AtomDestiny::Balance::VsLightDefenceCoefficient);
         break;
 
     case EDefenceType::Medium:
-        damage *= AtomDestiny::Balance::VsMediumDefenceCoefficient;
+        damage *= static_cast<float>(AtomDestiny::Balance::VsMediumDefenceCoefficient);
         break;
 
     case EDefenceType::Heavy:
-        damage *= AtomDestiny::Balance::VsHeavyDefenceCoefficient;
+        damage *= static_cast<float>(AtomDestiny::Balance::VsHeavyDefenceCoefficient);
         break;
 
     default:
@@ -152,13 +152,13 @@ void UParametersBase::RecalculateParameter(EObjectParameters parameter)
         LOG_WARNING(TEXT("[UObjectStateBase] Parameter is not available for recalculation"));
         return;
     }
-    
+
     switch (parameter)
     {
     case EObjectParameters::MaxHealth:
         {
             m_currentMaxHealth = CalculateParametersFromAll(m_maxHealth, parameter);
-            
+
             if (m_currentHealth == m_maxHealth)
                 m_currentHealth = m_currentMaxHealth;
             else
@@ -166,12 +166,12 @@ void UParametersBase::RecalculateParameter(EObjectParameters parameter)
         }
 
         break;
-    
+
     case EObjectParameters::Health:
         {
             std::vector<FParameterEnhancement>& parameters = GetParameterEnhancementList(parameter);
             const double value = InterpretParameterModifier(m_currentHealth, parameters.empty() ? FParameterEnhancement{} : parameters.front());
-            
+
             parameters.clear();
             m_currentHealth += value;
 
@@ -182,7 +182,7 @@ void UParametersBase::RecalculateParameter(EObjectParameters parameter)
 
             break;
         }
-    
+
     case EObjectParameters::Defence:
         m_currentDefence = CalculateParametersFromAll(m_defence, parameter);
         break;
